requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiomysql==0.2.0
aioredis==2.0.1
aiosignal==1.3.2
aiosmtplib==3.0.2
aiosqlite==0.20.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.5.2
async-timeout==5.0.1
asyncio-mqtt==0.16.2
attrs==25.3.0
bcrypt==4.0.0
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==44.0.0
dataclasses-json==0.6.7
dill==0.3.9
distro==1.9.0
docopt==0.6.2
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.12
filetype==1.2.0
frozenlist==1.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-genai==1.16.1
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.72.0rc1
grpcio-status==1.71.0
h11==0.14.0
hatchling==1.27.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
inquirerpy==0.3.4
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.61
langchain-google-genai==2.1.4
langchain-mcp-adapters==0.1.1
langchain-openai==0.3.18
langchain-text-splitters==0.3.8
langgraph==0.4.7
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.1
langgraph-sdk==0.1.70
langsmith==0.3.42
Mako==1.3.8
MarkupSafe==2.1.5
marshmallow==3.26.1
mcp==1.9.1
multidict==6.4.4
multiprocess==0.70.14
mypy_extensions==1.1.0
nestd==0.3.1
numpy==2.2.6
openai==1.58.1
orjson==3.9.15
ormsgpack==1.10.0
packaging==24.2
paho-mqtt==2.1.0
passlib==1.7.4
pathspec==0.12.1
pfzy==0.3.4
pluggy==1.6.0
prompt_toolkit==3.0.48
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.27.2
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
robyn==0.64.2
rsa==4.9
rustimport==1.3.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sse-starlette==2.3.5
starlette==0.46.2
stomp.py==8.2.0
tenacity==9.1.2
tiktoken==0.9.0
toml==0.10.2
tomli==2.2.1
tqdm==4.67.1
trove-classifiers==2025.5.9.12
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.4.0
uv==0.7.8
uvicorn==0.34.2
uvloop==0.19.0
volcengine-python-sdk==3.0.7
watchdog==4.0.1
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.1
xxhash==3.5.0
yarl==1.20.0
zipp==3.20.2
zstandard==0.23.0
